Ensure that older versions of git correctly import commits with empty committer name.
It seems that in older versions of git (1.7.1 from RHEL 6.3) fast-import won't
correctly sanitize committer field with empty username. This results in commit
objects with just one space between 'committer' and '<'. And newer versions of
git are producing plenty of warnings about those malformed commit objects.

One workaround which seems to be working with both git 1.7.1 and 1.8.0 is have
that extra space in fast-import stream. Strictly speaking this violates fast-import
spec, but seems to be working fine. Both 1.7.1 and 1.8.0 accept resulting stream
and produce correct commit objects with the same SHA.